requirements/base.txt
asyncio==0.4.1
pyserial==2.7
